[lint] fixed errors and added a few javadocs

+++ DogshowPro/myDogshow/src/main/java/dev/tnclark8012/apps/android/dogshow/util/Lists.java
@@ -17,7 +17,7 @@ public class Lists {
 		for(int i = 0; i < length; i++)
 		{
 			current = ints.get(i);
-			primitive[i] = (current!=null) ?  current.intValue() : defaultValue;
+			primitive[i] = (current!=null) ?  current : defaultValue;
 		}
 		return primitive;
 	}
